Refactored worker to only use 1 instance, refactored API call
