+++ src/components/NavItems/tools/SyntheticImageDetection/index.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { useLocation } from "react-router-dom";
 
@@ -13,22 +13,18 @@ import Switch from "@mui/material/Switch";
 
 import { Gradient } from "@mui/icons-material";
 
+import { useSyntheticImageDetection } from "@/components/NavItems/tools/SyntheticImageDetection/useSyntheticImageDetection";
 import { ROLES } from "@/constants/roles";
+import { resetSyntheticImageDetectionImage } from "@/redux/actions/tools/syntheticImageDetectionActions";
 import {
-  resetSyntheticImageDetectionImage,
-  setSyntheticImageDetectionLoading,
-  setSyntheticImageDetectionNearDuplicates,
-  setSyntheticImageDetectionResult,
-} from "@/redux/actions/tools/syntheticImageDetectionActions";
-import { isValidUrl } from "@Shared/Utils/URLUtils";
-import { preprocessFileUpload } from "@Shared/Utils/fileUtils";
-import axios from "axios";
+  preprocessFileUpload,
+  resizeImageWithWorker,
+} from "@Shared/Utils/fileUtils";
 import { i18nLoadNamespace } from "components/Shared/Languages/i18nLoadNamespace";
 import { setError } from "redux/reducers/errorReducer";
 
 import HeaderTool from "../../../Shared/HeaderTool/HeaderTool";
 import StringFileUploadField from "../../../Shared/StringFileUploadField";
-import { syntheticImageDetectionAlgorithms } from "./SyntheticImageDetectionAlgorithms";
 import SyntheticImageDetectionResults from "./syntheticImageDetectionResults";
 
 const SyntheticImageDetection = () => {
@@ -66,7 +62,11 @@ const SyntheticImageDetection = () => {
     UPLOAD: "local",
   };
 
-  const workerRef = useRef(null);
+  const { getSyntheticImageScores } = useSyntheticImageDetection({
+    dispatch,
+    keyword,
+    keywordWarning,
+  });
 
   useEffect(() => {
     if (urlParam && !input) {
@@ -80,190 +80,6 @@ const SyntheticImageDetection = () => {
       handleSubmit(input);
     }
   }, [url, input, result]);
-
-  useEffect(() => {
-    workerRef.current = new Worker(
-      new URL("@workers/resizeImageWorker", import.meta.url),
-    );
-
-    return () => {
-      workerRef.current.terminate();
-    };
-  }, []);
-
-  /**
-   *
-   * @param image
-   */
-  const resizeImageWithWorker = (image) => {
-    return new Promise((resolve, reject) => {
-      const workerInstance = new Worker(
-        new URL("@workers/resizeImageWorker", import.meta.url),
-      );
-      workerInstance.postMessage(image);
-
-      workerInstance.onerror = function (e) {
-        reject(e.error);
-      };
-
-      workerInstance.onmessage = function (e) {
-        // console.log(e);
-        resolve(e.data);
-      };
-    });
-  };
-
-  const getSyntheticImageScores = async (
-    url,
-    processURL,
-    dispatch,
-    type,
-    image,
-  ) => {
-    if (!processURL || (!url && !image)) {
-      return;
-    }
-
-    dispatch(setSyntheticImageDetectionLoading(true));
-    const modeURL = "images/";
-
-    const services = syntheticImageDetectionAlgorithms
-      .map((algorithm) => algorithm.apiServiceName)
-      .join(",");
-
-    const baseURL = process.env.REACT_APP_CAA_DEEPFAKE_URL;
-
-    const getUserFriendlyError = (error) => {
-      // Default error
-      if (!error) {
-        return keyword("synthetic_image_detection_error_generic");
-      }
-
-      if (
-        error.includes("Received status code 400") ||
-        error.includes("Cannot open image from")
-      )
-        return keyword("synthetic_image_detection_error_400");
-
-      return keyword("synthetic_image_detection_error_generic");
-    };
-
-    const handleError = (e) => {
-      dispatch(setError(e));
-      dispatch(setSyntheticImageDetectionLoading(false));
-    };
-
-    if (!isValidUrl(url) && !image) {
-      handleError(keywordWarning("error_invalid_url"));
-      return;
-    }
-
-    let res;
-    const bodyFormData = new FormData();
-
-    try {
-      switch (type) {
-        case IMAGE_FROM.UPLOAD:
-          bodyFormData.append("file", image);
-          res = await axios.post(baseURL + modeURL + "jobs", bodyFormData, {
-            method: "post",
-            params: { services: services, search_similar: true },
-            headers: {
-              "Content-Type": "multipart/form-data",
-            },
-          });
-          break;
-
-        default:
-          res = await axios.post(baseURL + modeURL + "jobs", null, {
-            params: { url: url, services: services, search_similar: true },
-          });
-          break;
-      }
-    } catch (error) {
-      const processedError = getUserFriendlyError(
-        error?.response?.data?.message ?? "error_" + error.status,
-      );
-      handleError(processedError);
-    }
-
-    const getResult = async (id) => {
-      let response;
-      try {
-        response = await axios.get(baseURL + modeURL + "reports/" + id);
-      } catch (error) {
-        handleError("error_" + error.status);
-      }
-
-      if (response && response.data != null) {
-        dispatch(
-          setSyntheticImageDetectionResult({
-            url: image ? URL.createObjectURL(image) : url,
-            result: response.data,
-          }),
-        );
-      }
-
-      if (
-        response &&
-        response.data &&
-        response.data.similar_images &&
-        response.data.similar_images.completed
-      ) {
-        let imgSimilarRes;
-
-        try {
-          imgSimilarRes = await axios.get(baseURL + modeURL + "similar/" + id);
-        } catch (error) {
-          handleError("error_" + error.status);
-        }
-
-        //console.log(imgSimilarRes.data);
-
-        if (
-          !imgSimilarRes.data ||
-          !imgSimilarRes.data.similar_media ||
-          !Array.isArray(imgSimilarRes.data.similar_media) ||
-          imgSimilarRes.data.similar_media.length === 0
-        ) {
-          dispatch(setSyntheticImageDetectionNearDuplicates(null));
-        }
-
-        dispatch(setSyntheticImageDetectionNearDuplicates(imgSimilarRes.data));
-      }
-    };
-
-    const waitUntilFinish = async (id) => {
-      let response;
-
-      try {
-        response = await axios.get(baseURL + modeURL + "jobs/" + id);
-      } catch (error) {
-        handleError("error_" + error.status);
-      }
-
-      if (response && response.data && response.data.status === "PROCESSING") {
-        await sleep(waitUntilFinish, id);
-      } else if (
-        response &&
-        response.data &&
-        response.data.status === "COMPLETED"
-      ) {
-        await getResult(id);
-      } else {
-        handleError("error_" + response.data.status);
-      }
-    };
-
-    if (!res || !res.data) return;
-    await waitUntilFinish(res.data.id);
-  };
-
-  function sleep(fn, param) {
-    return new Promise((resolve) => {
-      setTimeout(() => resolve(fn(param)), 3000);
-    });
-  }
 
   const resetState = () => {
     setInput("");
@@ -308,19 +124,15 @@ const SyntheticImageDetection = () => {
     }
 
     dispatch(resetSyntheticImageDetectionImage());
+
     const urlInput = url ? url : input;
+
     const type =
       urlInput && typeof urlInput === "string"
         ? IMAGE_FROM.URL
         : IMAGE_FROM.UPLOAD;
 
-    await getSyntheticImageScores(
-      urlInput,
-      true,
-      dispatch,
-      type,
-      processedFile,
-    );
+    await getSyntheticImageScores(urlInput, true, type, processedFile);
   };
 
   useEffect(() => {
